+++ changelog.md
@@ -97,6 +97,7 @@ Last change: 07-MAY-2026 CEST
 - `picasso.gausslq.initial_parameters_gpufit` and `picasso.gaussmle.mean_filter` will become private in v0.11.0
 - `picasso.localize` functions: `local_maxima`, `gradient_at`, `net_gradient` will become private in v0.11.0. Functions `fit` and `fit_async` will be removed entirely
 - `picasso.postprocess` functions: `index_blocks_shape`, `n_block_locs_at`, `next_frame_neighbor_distance_histogram`, `get_link_groups` and `link_loc_groups` will become private in v0.11.0
+- `picasso.spinna` functions: `find_target_counts`, `get_structures_permutation`, `targets_from_structures`
 
 ## 0.9.10